backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "AI Course Generator Backend API"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "pydantic",
    "passlib[bcrypt]",
    "python-jose[cryptography]",
    "python-multipart",
    "pymongo",
    "python-dotenv",
    "PyJWT",
    "langgraph",
    "langchain-core",
    "langchain-openai",
    "openai",
    "youtube-search-python",
    "langchain-groq>=1.1.2",
    "langchain-google-genai>=4.2.1",
]
